Derp

+++ Web/Place.py
@@ -27,7 +27,7 @@ def set_shape(shape):
         for y in range(15):
             if [x,y] in data:
                 if active_grid[x,y] is None:
-                    active_grid[x,y] = np.zeros((32, 32, 3), dtype=numpy.int8)
+                    active_grid[x,y] = np.zeros((32, 32, 3), dtype=int)
             else:
                 active_grid[x,y] = None
     return "Good"
